+++ src/resample/ResampleScrambleVarMonths.jl
@@ -1,10 +1,12 @@
 # scramblevar.jl - Functions to resample VarCPIBase objects
 
+import Random: AbstractRNG
+
 # This is the best version, it requires creating copies of the vectors of the same
 # months, for each basic expenditure. A more efficient version is presented below
 # 475.600 μs (2618 allocations: 613.20 KiB)
 
-# function scramblevar!(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG) 
+# function scramblevar!(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)
 #     for i in 1:12
 #         # fill every column with random values from the same periods (t and t+12)
 #         for j in 1:size(vmat, 2)
@@ -14,43 +16,45 @@
 # end
 
 
-# function scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG) 
+# function scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)
 #     scrambled_mat = copy(vmat)
 #     scramblevar!(scrambled_mat, rng)
 #     scrambled_mat
 # end
 
-# First version with column resampling 
-# function scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG) 
+# First version with column resampling
+# function scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)
 #     periods, n = size(vmat)
-#     # Matrix of resampled values 
-#     v_sc = similar(vmat) 
+#     # Matrix of resampled values
+#     v_sc = similar(vmat)
 #     for i in 1:min(periods, 12)
 #         v_month = vmat[i:12:periods, :]
 #         periods_month = size(v_month, 1)
-#         for g in 1:n 
+#         for g in 1:n
 #             v_month[:, g] = rand(rng, v_month[:, g], periods_month)
-#         end       
+#         end
 #         # Assign values of the same months
 #         v_sc[i:12:periods, :] = v_month
 #     end
 #     v_sc
 # end
 
-# Memory-optimized version 
+# Memory-optimized version
 # 420.100 μs (2 allocations: 204.45 KiB)
-function scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG) 
-    periods, n = size(vmat)
+function scramblevar(vmat::AbstractMatrix, rng::AbstractRNG = Random.GLOBAL_RNG, sample_periods::Int = size(vmat, 1))
+    # periods = number of periods to resample
+    # Actual available periods in vmat
+    actual_periods, n = size(vmat)
 
-    # Matrix of resampled values 
-    v_sc = similar(vmat) 
+    # Matrix of resampled values
+    v_sc = similar(vmat, sample_periods, n)
 
     # For each month and each basic expenditure, randomly take from the same
     # months of vmat and fill v_sc (v_scrambled)
-    for i in 1:min(periods, 12), g in 1:n 
-        Random.rand!(rng, view(v_sc, i:12:periods, g), view(vmat, i:12:periods, g))        
-    end    
-    v_sc
+    for i in 1:min(sample_periods, 12), g in 1:n
+        Random.rand!(rng, view(v_sc, i:12:sample_periods, g), view(vmat, i:12:actual_periods, g))
+    end
+    return v_sc
 end
 
 
@@ -72,11 +76,11 @@ get_param_function(::ResampleScrambleVarMonths) = param_scramblevar_fn
 
 # Define how to resample matrices with time series in the columns.
 # Uses the internal function `scramblevar`.
-function (resamplefn::ResampleScrambleVarMonths)(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG) 
-    scramblevar(vmat, rng)
-end 
+function (resamplefn::ResampleScrambleVarMonths)(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)
+    return scramblevar(vmat, rng)
+end
 
-# Define the name and tag of the resampling method 
+# Define the name and tag of the resampling method
 method_name(resamplefn::ResampleScrambleVarMonths) = "IID bootstrap by months of occurrence"
 method_tag(resamplefn::ResampleScrambleVarMonths) = "SVM"
 
@@ -90,31 +94,29 @@ bootstrap resampling methodology by the same calendar months. Returns a
 months of occurrence (also called population monthly price changes).
 
 """
-function param_scramblevar_fn(base::VarCPIBase)
-
-    # Obtener matriz de promedios mensuales
-    month_mat = monthavg(base.v)
-
-    # Conformar base de variaciones intermensuales promedio
-    VarCPIBase(month_mat, base.w, base.dates, base.baseindex)
+function param_scramblevar_fn(base::VarCPIBase, sample_periods::Int = periods(base))
+    # Get the matrix of average monthly price changes
+    month_mat = monthavg(base.v, sample_periods)
+    # Form the VarCPIbase with monthly averages 
+    return VarCPIBase(month_mat, base.w, base.dates, base.baseindex)
 end
 
 function param_scramblevar_fn(cs::CountryStructure)
     pob_base = map(param_scramblevar_fn, cs.base)
-    getunionalltype(cs)(pob_base)
+    return getunionalltype(cs)(pob_base)
 end
 
 # Obtener variaciones intermensuales promedio de los mismos meses de ocurrencia.
-# Se remuestrean `numobsresample` observaciones de las series de tiempo en las
-# columnas de `vmat`. 
-function monthavg(vmat, numobsresample = size(vmat, 1))
-    # Crear la matriz de promedios 
+# Se remuestrean `sample_periods` observaciones de las series de tiempo en las
+# columnas de `vmat`.
+function monthavg(vmat::AbstractMatrix, sample_periods::Int = size(vmat, 1))
+    # Crear la matriz de promedios
     cols = size(vmat, 2)
-    avgmat = Matrix{eltype(vmat)}(undef, numobsresample, cols)
-    
-    # Llenar la matriz de promedios con los promedios de cada mes 
+    avgmat = similar(vmat, sample_periods, cols)
+
+    # Llenar la matriz de promedios con los promedios de cada mes
     for i in 1:12
-        avgmat[i:12:end, :] .= mean(vmat[i:12:end, :], dims=1)
+        avgmat[i:12:end, :] .= mean(vmat[i:12:end, :], dims = 1)
     end
     return avgmat
 end