+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] - 2025-10
 ### Added 
 - Type `CPIVarietyMatchDistribution` to resample as in the B-TIMA extension methodology. It uses two arrays of monthly price changes, a prior and an actual empirical distribution. It can be configured to sample from the prior distribution, the actual  distribution or a synthtetic distribution that gives different weights according to the mean of the actual observations.
+- Resampling function `ResampleExtendedSVM` to sample CPI datasets as in the original B-TIMA methodology, but allowing for more periods in the sampling. 
 - Resampling function `ResampleSynthetic` to implement an interface to resample `VarCPIBase` objects.
 - Resampling function `ResampleIdentity` to resample `CountryStructure`s and `VarCPIBase` objects. 
 - Resampling function `ResampleMixture` to implement an ensemble of samplers for the `VarCPIBase` components. This allows setting mixed resampling schemes for different CPI datasets. For example, in a `CountryStructure` with two `VarCPIBase`s, not resampling the first dataset (`ResampleIdentity`) and using the same calendar months for the second CPI dataset (`ResampleScrambleVarMonths`).